CORE: Introduce NodeContext class from upstream
This is used to contain/pass some frequently used global variables around instead of passing all the variables around seperately.
We currentlyonly place the scheduler on this, as our codebase is very different from upstream, especially in terms of init behaviour (where this is heavily featured) we won't use this entirely the same as them.
However we use the same structure where it makes sense to keep our codebases a bit more aligned.

Bring our codebase more inline with upstream to ease backporting of other changes